pyproject.toml
[project]
name = "idap-data-pipelines-aus-refdatapipeline-skilltraining"
version = "0.1.0"
description = ""
authors = [
    {name = "wacheng",email = "wacheng@corelogic.com"}
]
readme = "README.md"
requires-python = "^3.11"
dependencies = [
    "requests (>=2.32.3,<3.0.0)",
    "aiohttp (>=3.11.18,<4.0.0)",
    "pytest-asyncio (>=0.26.0,<0.27.0)",
    "dotenv (>=0.9.9,<0.10.0)",
    "pandas (>=2.2.3,<3.0.0)",
    "duckdb (>=1.2.2,<2.0.0)",
    "google-auth (>=2.39.0,<3.0.0)",
    "google-cloud-bigquery (>=3.31.0,<4.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.5"

